Remove incorrect parentheses for AI-deck UART1 and UART2 connections
As indicated in the description above the deck pin allocation table, parentheses should indicate that pins are physically unconnected. However, UART1 and UART2 on the AI-deck are not physically unconnected. They may not be required for most funcitonalities (although UART2 has become important by defaultwith the introduction of CPX), but that should be indicated otherwise, if desired.