Add warning about missing autotiler file to documentation
Added a warning to AI-deck getting started guide regarding the unavailability of the required autotiler file for neural network deployment with gap_sdk due to the GreenWaves Technologies website being down. The warning informs users that he autotiler will not work unless they already have the file and provides a link to the GitHub discussions announcement with more details and workarounds.